// shifthappens/types/enums.ts
// Enums
export enum RequestType {
    Exclude = "exclude",
    Prefer = "prefer",
  }
  
  export enum AssignmentType {
    Auto = "auto",
    Admin = "admin",
  }
  
  export enum MultiplierType {
    User = "user",
    Shift = "shift",
  }
  
  export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other",
  }